\name{eigendog-internal}
\alias{bitSetOfRLE}
\alias{checkConsistencyTree}
\alias{checkConsistencyTrees}
\alias{confusion}
\alias{feature}
\alias{indexByElement}
\alias{model}
\alias{predictTree}
\alias{predictTrees}
\alias{probability}
\alias{trainResultOfJson}
\alias{trainResultV1}
\alias{tree}
\title{Internal eigendog Functions}
\description{
  Internal eigendog functions
}
\details{
  These are inernal functions, and as such, are more likely to change
  across revisions.
}
\keyword{ internal }